+++ app/models/reserved_domain.rb
@@ -30,6 +30,8 @@ class ReservedDomain < ActiveRecord::Base
   end
 
   def generate_data
+    return if Domain.where(name: name).any?
+
     @json = generate_json
     @body = generate_body
     update_whois_server
@@ -57,6 +59,8 @@ class ReservedDomain < ActiveRecord::Base
   end
 
   def remove_data
+    return if Domain.where(name: name).any?
+
     Whois::Record.where(name: name).delete_all
   end
